+++ uvicorn/_impl/asyncio/loops/uvloop.py
@@ -1,0 +1,7 @@
+import asyncio
+
+import uvloop
+
+
+def uvloop_setup():
+    asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
