+++ uvicorn/loops/asyncio.py
@@ -1,18 +1,6 @@
-import asyncio
-import platform
-import selectors
-import sys
+"""
+Backwards compatibility shim.
+"""
+from .._impl.asyncio.loops.asyncio import asyncio_setup
 
-
-def asyncio_setup():
-    if (
-        sys.version_info.major >= 3
-        and sys.version_info.minor >= 8
-        and platform.system() == "Windows"
-    ):
-        selector = selectors.SelectSelector()
-        loop = asyncio.SelectorEventLoop(selector)
-        asyncio.set_event_loop(loop)
-    else:
-        loop = asyncio.new_event_loop()
-        asyncio.set_event_loop(loop)
+__all__ = ["asyncio_setup"]
