+++ uvicorn/loops/auto.py
@@ -1,11 +1,6 @@
-def auto_loop_setup():
-    try:
-        import uvloop  # noqa
-    except ImportError:  # pragma: no cover
-        from uvicorn.loops.asyncio import asyncio_setup as loop_setup
+"""
+Backwards compatibility shim.
+"""
+from .._impl.asyncio.loops.auto import auto_loop_setup
 
-        loop_setup()
-    else:
-        from uvicorn.loops.uvloop import uvloop_setup
-
-        uvloop_setup()
+__all__ = ["auto_loop_setup"]
