+++ uvicorn/protocols/http/auto.py
@@ -1,10 +1,6 @@
-try:
-    import httptools  # noqa
-except ImportError:  # pragma: no cover
-    from uvicorn.protocols.http.h11_impl import H11Protocol
+"""
+Backwards compatibility shim.
+"""
+from uvicorn._impl.asyncio.protocols.http.auto import AutoHTTPProtocol
 
-    AutoHTTPProtocol = H11Protocol
-else:
-    from uvicorn.protocols.http.httptools_impl import HttpToolsProtocol
-
-    AutoHTTPProtocol = HttpToolsProtocol
+__all__ = ["AutoHTTPProtocol"]
